+++ pyincore_viz/geoutil.py
@@ -18,7 +18,6 @@ import PIL
 import numpy as np
 import random
 import json
-import contextily as ctx
 
 from deprecated.sphinx import deprecated
 from matplotlib import cm
@@ -31,7 +30,7 @@ from pyincore.dataservice import DataService
 from pyincore.hazardservice import HazardService
 from pyincore import Dataset
 from pyincore import NetworkDataset
-from pyincore_viz import globals
+from pyincore_viz import globals as pyincore_viz_globals
 from base64 import b64encode
 from io import BytesIO
 from pyincore_viz.plotutil import PlotUtil
@@ -39,7 +38,7 @@ from pyincore_viz.tabledatasetlistmap import TableDatasetListMap as table_list_m
 from pyincore_viz.helpers.common import get_period_and_demand_from_str, get_demands_for_dataset_hazards
 from branca.colormap import linear
 
-logger = globals.LOGGER
+logger = pyincore_viz_globals.LOGGER
 
 
 class GeoUtil:
@@ -72,7 +71,7 @@ class GeoUtil:
         Args:
             gdf (obj): Geopandas DataFrame object.
             column (obj): A column name of gdf to be plot.
-            hazard_id (str): A raster hazard dataset id to overlay, such as tif or png dataset
+            raster (str): A raster hazard dataset id to overlay, such as tif or png dataset
             category (bool): Turn on/off category option.
             basemap (bool): Turn on/off base map (e.g. openstreetmap).
             source(obj): source of the Map to be used. examples, ctx.providers.OpenStreetMap.Mapnik (default),
@@ -399,7 +398,7 @@ class GeoUtil:
         return m
 
     @staticmethod
-    def get_wms_map(datasets: list, wms_url=globals.INCORE_GEOSERVER_WMS_URL, layer_check=False):
+    def get_wms_map(datasets: list, wms_url=pyincore_viz_globals.INCORE_GEOSERVER_WMS_URL, layer_check=False):
         """Get a map with WMS layers from list of datasets.
 
         Args:
@@ -465,7 +464,7 @@ class GeoUtil:
         return m
 
     @staticmethod
-    def get_gdf_wms_map(datasets, wms_datasets, wms_url=globals.INCORE_GEOSERVER_WMS_URL):
+    def get_gdf_wms_map(datasets, wms_datasets, wms_url=pyincore_viz_globals.INCORE_GEOSERVER_WMS_URL):
         """Get a map with WMS layers from list of datasets for geopandas and list of datasets for WMS.
 
         Args:
@@ -477,7 +476,6 @@ class GeoUtil:
             obj: An ipyleaflet Map.
 
         """
-
         # TODO: how to add a style for each WMS layers (pre-defined styules on WMS server) and gdf layers
 
         # (min_lat, min_lon, max_lat, max_lon)
@@ -620,14 +618,14 @@ class GeoUtil:
         return joined_gdf
 
     @staticmethod
-    def plot_table_dataset_list_from_single_source(client, dataset_list=list, column=str, in_source_dataset_id=None):
+    def plot_table_dataset_list_from_single_source(client, dataset_list, column, in_source_dataset_id=None):
         """Creates map window with a list of table dataset and source dataset.
 
             Args:
                 client (obj): pyincore service Client Object.
                 dataset_list (list): list of table dataset.
                 column (str): column name to be plot.
-                in_source_dataset_id (str): source dataset id, the dafault is None.
+                in_source_dataset_id (str): source dataset id, the default is None.
 
             Returns:
                 obj: An ipyleaflet Map, GeoUtil.map (ipyleaflet.Map).
@@ -671,7 +669,7 @@ class GeoUtil:
         Args:
             dataset_list (list): list of table dataset.
             column (str): column name to be plot.
-            source_dataset (str): source dataset id, default is None.
+            in_source_dataset_id (str): source dataset id, default is None.
 
         Returns:
             obj: Pandas dataframe with all dataset joined together with guid.
@@ -1370,7 +1368,7 @@ class GeoUtil:
                 field_list (list): A list of fields in the dataset.
                         The order of the list should be matched with the order of dataset list
                 zoom_level (int): Zoom level
-                
+
             Returns:
                 obj: An ipyleaflet map.