+++ docs/source/conf.py
@@ -31,9 +31,9 @@ project = 'pyIncore-viz'
 author = 'Yong Wook Kim'
 
 # The short X.Y version
-version = '1.9'
+version = '1.10'
 # The full version, including alpha/beta/rc tags
-release = '1.9.0'
+release = '1.10.0'
 
 # -- General configuration ---------------------------------------------------
 
@@ -90,7 +90,9 @@ pygments_style = 'sphinx'
 # See also:
 # http://www.sphinx-doc.org/en/stable/ext/autodoc.html#confval-autodoc_mock_importshttps://github.com/sphinx-doc/sphinx/issues/4182
 
-autodoc_mock_imports = ['pytest', 'rasterstats']
+autodoc_mock_imports = ['pyincore_viz', 'pytest', 'rasterstats']
+
+
 
 # This value selects what content will be inserted into the main body of an autoclass directive.
 # The possible values are: