+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Refactor tornado and eq visualization due to hazard datasets added to model [#154](https://github.com/IN-CORE/pyincore-viz/issues/154)
 
+### Fixed
+- Incorrect documentation rendering has been fixed [#158](https://github.com/IN-CORE/pyincore-viz/issues/158)
+
 ## [1.9.0] - 2023-12-13
 ### Added
 - Local hazard visualization [#143](https://github.com/IN-CORE/pyincore-viz/issues/143)