+++ CHANGELOG.md
@@ -6,8 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Plot method for local hazard [#153](https://github.com/IN-CORE/pyincore-viz/issues/153)
+
 ### Changed
 - Refactor tornado and eq visualization due to hazard datasets added to model [#154](https://github.com/IN-CORE/pyincore-viz/issues/154)
+
 
 ## [1.9.0] - 2023-12-13
 ### Added