+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Refactor tornado and eq visualization due to hazard datasets added to model [#154](https://github.com/IN-CORE/pyincore-viz/issues/154)
 
+### Fixed
+- Incorrect documentation rendering has been fixed [#158](https://github.com/IN-CORE/pyincore-viz/issues/158)
 
 ## [1.9.0] - 2023-12-13
 ### Added